fix(soft-delete-purge-sweeper): actually purge credentials 168h after soft-delete
Closes production-readiness-audit finding #6 (Blocker), per ADR-073.
deleted_at was write-only -- nothing ever read it, so soft-deleted
projects' ECIES-encrypted DSNs/credentials persisted forever,
contradicting the admin-UI's own explicit 168h purge promise.

New sweepers/soft_delete_purge_sweeper.rs mirrors cap_usage_refresher.rs's
shape (SystemDb-only, advisory-lock-guarded interval loop) -- no
customer-DB connection needed, since the target columns live on the
projects row itself. Single UPDATE nulls ecies_encrypted_dsn/
backend_pg_dsn_enc/agent_tls_bundle_enc for rows soft-deleted past the
grace window; the nulled-columns condition is its own idempotency
guard, no new column needed. New config: EMBYR_SOFT_DELETE_SWEEP_INTERVAL_SECS
(default 3600), EMBYR_SOFT_DELETE_GRACE_DAYS (default 7, matching the
admin-UI's unchanged 168h promise).

DELIVER found and fixed a real test-fixture bug (not production code,
not an assertion): make_interval(days => $5) needed an explicit ::int
cast since $5 was otherwise inferred float8 from a sibling bind-param
cast on the same line.

7/7 acceptance scenarios pass reliably (2 independent runs by DELIVER,
2 more independent runs by the orchestrator). Sibling regression guard
(customer_db_transaction_sweeper_us01/us02) unmodified and green.

+++ tests/soft_delete_purge_sweeper/acceptance/us01_purge_credentials_after_grace_window.rs
@@ -115,7 +115,7 @@ async fn insert_project(
          (id, account_id, backend_mode, api_key_hash_current, status, name, \
           deleted_at, ecies_encrypted_dsn, backend_pg_dsn_enc, agent_tls_bundle_enc) \
          VALUES ($1, $2, $3, 'unused-hash', $4, $1, \
-          CASE WHEN $5::float8 IS NULL THEN NULL ELSE now() - make_interval(days => $5) END, \
+          CASE WHEN $5::float8 IS NULL THEN NULL ELSE now() - make_interval(days => $5::int) END, \
           $6, $7, $8)",
     )
     .bind(project_id)
